Avoid getting a threadlocal twice.
currentHub() is a hotspot method and ThreadLocal.get() involves a map
lookup (the more thread locals there are in the current thread, the
more expensive that is and the number of thread locals in the current
thread is out of our control).

Avoiding the extraneous ThreadLocal.get() is quite easy here so we
should not do it.